# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BookingDetails } from './bookingDetails';
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  threadId: number;
  /** @nullable */
  createdByUserId?: number | null;
  /** @nullable */
  approverUserId?: number | null;
  /** @nullable */
  approverPhoneNumber?: string | null;
  title: string;
  details: BookingDetails;
  status: BookingStatus;
  /** @nullable */
  provider?: string | null;
  /** @nullable */
  providerBookingId?: string | null;
  createdAt: Date;
  updatedAt: Date;
  /** @nullable */
  decidedAt?: Date | null;
}