# lib/api-zod/src/generated/types/bookingStatus.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  drafted: 'drafted',
  pending_approval: 'pending_approval',
  approved: 'approved',
  rejected: 'rejected',
  confirmed: 'confirmed',
} as const;