# lib/api-zod/src/generated/types/onboardingStatus.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type OnboardingStatus = typeof OnboardingStatus[keyof typeof OnboardingStatus];


export const OnboardingStatus = {
  not_started: 'not_started',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;