# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  userId: number;
  /** @nullable */
  budget?: string | null;
  /** @nullable */
  dietaryNeeds?: string | null;
  preferences: string[];
  pastChoices: string[];
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}